Move AddTable to sleeper.clients.status.update
