+++ java/clients/src/main/java/sleeper/clients/status/update/AddTable.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.clients.util;
+package sleeper.clients.status.update;
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
