+++ java/clients/src/test/java/sleeper/clients/status/update/AddTableIT.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.clients.util;
+package sleeper.clients.status.update;
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
