+++ scripts/deploy/addTable.sh
@@ -52,4 +52,4 @@ sed \
 echo "-------------------------------------------------------"
 echo "Adding table"
 echo "-------------------------------------------------------"
-java -cp "${JAR_DIR}/clients-${VERSION}-utility.jar" sleeper.clients.util.AddTable "${INSTANCE_ID}" "${TABLE_PROPERTIES}"
+java -cp "${JAR_DIR}/clients-${VERSION}-utility.jar" sleeper.clients.status.update.AddTable "${INSTANCE_ID}" "${TABLE_PROPERTIES}"
